Struct google_surveys2::api::SurveyQuestion

pub struct SurveyQuestion {
    pub answer_order: Option<String>,
    pub answers: Option<Vec<String>>,
    pub has_other: Option<bool>,
    pub high_value_label: Option<String>,
    pub images: Option<Vec<SurveyQuestionImage>>,
    pub last_answer_position_pinned: Option<bool>,
    pub low_value_label: Option<String>,
    pub must_pick_suggestion: Option<bool>,
    pub num_stars: Option<String>,
    pub open_text_placeholder: Option<String>,
    pub open_text_suggestions: Option<Vec<String>>,
    pub question: Option<String>,
    pub sentiment_text: Option<String>,
    pub single_line_response: Option<bool>,
    pub threshold_answers: Option<Vec<String>>,
    pub type_: Option<String>,
    pub unit_of_measurement_label: Option<String>,
    pub video_id: Option<String>,
}

Message defining the question specifications.

This type is not used in any activity, and only used as part of another schema.

Fields
answer_order: Option<String>

The randomization option for multiple choice and multi-select questions. If not specified, this option defaults to randomize.

answers: Option<Vec<String>>

Required list of answer options for a question.

has_other: Option<bool>

Option to allow open-ended text box for Single Answer and Multiple Answer question types. This can be used with SINGLE_ANSWER, SINGLE_ANSWER_WITH_IMAGE, MULTIPLE_ANSWERS, and MULTIPLE_ANSWERS_WITH_IMAGE question types.

high_value_label: Option<String>

For rating questions, the text for the higher end of the scale, such as 'Best'. For numeric questions, a string representing a floating-point that is the maximum allowed number for a response.

images: Option<Vec<SurveyQuestionImage>>

no description provided

last_answer_position_pinned: Option<bool>

Currently only support pinning an answer option to the last position.

low_value_label: Option<String>

For rating questions, the text for the lower end of the scale, such as 'Worst'. For numeric questions, a string representing a floating-point that is the minimum allowed number for a response.

must_pick_suggestion: Option<bool>

Option to force the user to pick one of the open text suggestions. This requires that suggestions are provided for this question.

num_stars: Option<String>

Number of stars to use for ratings questions.

open_text_placeholder: Option<String>

Placeholder text for an open text question.

open_text_suggestions: Option<Vec<String>>

A list of suggested answers for open text question auto-complete. This is only valid if single_line_response is true.

question: Option<String>

Required question text shown to the respondent.

sentiment_text: Option<String>

Used by the Rating Scale with Text question type. This text goes along with the question field that is presented to the respondent, and is the actual text that the respondent is asked to rate.

single_line_response: Option<bool>

Option to allow multiple line open text responses instead of a single line response. Note that we don't show auto-complete suggestions with multiple line responses.

threshold_answers: Option<Vec<String>>

The threshold/screener answer options, which will screen a user into the rest of the survey. These will be a subset of the answer option strings.

type_: Option<String>

Required field defining the question type. For details about configuring different type of questions, consult the question configuration guide.

unit_of_measurement_label: Option<String>

Optional unit of measurement for display (for example: hours, people, miles).

video_id: Option<String>

The YouTube video ID to be show in video questions.